review comments for simplying usage of page response headers
